Move asset-manifest modules into config namespace
