+++ app/config/asset-manifest.js
@@ -1,8 +1,8 @@
-import environment from './config/environment';
+import environment from './environment';
 
 const modulePrefix = environment.modulePrefix;
-const metaName = `${modulePrefix}/asset-manifest`;
-const nodeName = `${modulePrefix}/node-asset-manifest`;
+const metaName = `${modulePrefix}/config/asset-manifest`;
+const nodeName = `${modulePrefix}/config/node-asset-manifest`;
 
 let config = {};
