+++ lib/manifest-generator.js
@@ -21,7 +21,7 @@ var ManifestGenerator = Addon.extend({
     var app = findHost(this);
     var options = app && app.options && app.options.assetLoader;
     if (type === 'head-footer' && !(options && options.noManifest)) {
-      var metaName = config.modulePrefix + '/asset-manifest';
+      var metaName = config.modulePrefix + '/config/asset-manifest';
       return '<meta name="' + metaName + '" content="%GENERATED_ASSET_MANIFEST%" />';
     }
   },
