+++ lib/meta-handler.js
@@ -2,7 +2,7 @@
  * Replace the manifest meta tag with an updated version of the manifest.
  * We do this in both the app's and test's index.html file.
  */
-var regex = /<meta name="([^"]*\/asset-manifest)" content=[^>]*>/;
+var regex = /<meta name="([^"]*\/config\/asset-manifest)" content=[^>]*>/;
 
 function escaper(sourceJSON) {
   return escape(JSON.stringify(sourceJSON));
