+++ lib/utils/node-module-template.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-define('APP_NAME/node-asset-manifest', function() {
+define('APP_NAME/config/node-asset-manifest', function() {
   return {
     default: ASSET_MANIFEST
   };
