+++ node-tests/fixtures/generator-test/extra.html
@@ -13,7 +13,7 @@
     <link rel="stylesheet" href="/assets/vendor.css">
     <link rel="stylesheet" href="/assets/dummy.css">
 
-    <meta name="dummy/asset-manifest" content="" />
+    <meta name="dummy/config/asset-manifest" content="" />
   </head>
   <body>
