+++ node-tests/fixtures/insertion-test/tests/index.html
@@ -15,7 +15,7 @@
     <link rel="stylesheet" href="/assets/dummy.css">
     <link rel="stylesheet" href="/assets/test-support.css">
 
-    <meta name="dummy/asset-manifest" content="" />
+    <meta name="dummy/config/asset-manifest" content="" />
 
   </head>
   <body>
