+++ tests/unit/asset-manifest-test.js
@@ -5,25 +5,25 @@ import { module, test } from 'qunit';
 module('Unit | asset-manifest', {
   beforeEach() {
     resetModules();
-    this.originalNodeModule = require.entries['dummy/node-asset-manifest'];
+    this.originalNodeModule = require.entries['dummy/config/node-asset-manifest'];
   },
 
   afterEach() {
-    require.entries['dummy/node-asset-manifest'] = this.originalNodeModule;
+    require.entries['dummy/config/node-asset-manifest'] = this.originalNodeModule;
     resetModules();
   }
 });
 
 function resetModules() {
-  require.unsee('dummy/node-asset-manifest');
-  require.unsee('dummy/asset-manifest');
+  require.unsee('dummy/config/node-asset-manifest');
+  require.unsee('dummy/config/asset-manifest');
 }
 
 test('node-asset-manifest is generated properly', function(assert) {
-  const nodeManifest = require('dummy/node-asset-manifest').default;
-  delete require.entries['dummy/node-asset-manifest'];
+  const nodeManifest = require('dummy/config/node-asset-manifest').default;
+  delete require.entries['dummy/config/node-asset-manifest'];
 
-  const manifest = require('dummy/asset-manifest').default;
+  const manifest = require('dummy/config/asset-manifest').default;
 
   assert.notStrictEqual(nodeManifest, manifest);
   assert.deepEqual(nodeManifest, manifest);
@@ -31,27 +31,27 @@ test('node-asset-manifest is generated properly', function(assert) {
 
 test('loads the node-asset-manifest if present', function(assert) {
   const replacementModule = {};
-  define('dummy/node-asset-manifest', () => ({ default: replacementModule}));
+  define('dummy/config/node-asset-manifest', () => ({ default: replacementModule}));
 
-  assert.strictEqual(require('dummy/asset-manifest').default, replacementModule);
+  assert.strictEqual(require('dummy/config/asset-manifest').default, replacementModule);
 });
 
 test('loads the manifest from the meta tag if available', function(assert) {
-  delete require.entries['dummy/node-asset-manifest'];
+  delete require.entries['dummy/config/node-asset-manifest'];
 
-  const meta = document.querySelector('meta[name="dummy/asset-manifest"]');
+  const meta = document.querySelector('meta[name="dummy/config/asset-manifest"]');
   const metaContent = meta.getAttribute('content');
   meta.setAttribute('content', '{"derp":"herp"}');
-  assert.deepEqual(require('dummy/asset-manifest').default, { derp: 'herp' });
+  assert.deepEqual(require('dummy/config/asset-manifest').default, { derp: 'herp' });
   meta.setAttribute('content', metaContent);
 });
 
 test('throws an error if unable to load the manifest', function(assert) {
-  delete require.entries['dummy/node-asset-manifest'];
+  delete require.entries['dummy/config/node-asset-manifest'];
 
-  const meta = document.querySelector('meta[name="dummy/asset-manifest"]');
+  const meta = document.querySelector('meta[name="dummy/config/asset-manifest"]');
   const metaContent = meta.getAttribute('content');
   meta.setAttribute('content', 'herp');
-  assert.throws(() => assert.deepEqual(require('dummy/asset-manifest').default, {}));
+  assert.throws(() => assert.deepEqual(require('dummy/config/asset-manifest').default, {}));
   meta.setAttribute('content', metaContent);
 });
